refactor: layout 분리 및 outlet 적용
