+++ src/components/common/header/GNB.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Logo from '../../Logo';
+import Logo from '../Logo';
 import styled from 'styled-components';
 import Profile from './Profile';
