+++ src/layout/MainLayout.jsx
@@ -1,0 +1,16 @@
+import GNB from 'components/common/header/GNB';
+import Footer from 'components/common/footer/Footer';
+import { Outlet } from 'react-router-dom';
+
+const MainLayout = () => {
+  return (
+    <>
+      <GNB />
+      {/* 콘텐츠 영역 : 페이지마다 달라지는 영역 */}
+      <Outlet />
+      <Footer />
+    </>
+  );
+};
+
+export default MainLayout;
