+++ src/utils/constants/routes.js
@@ -4,6 +4,8 @@ const routes = {
   signin: '/signin',
   privacy: '/privacy',
   faq: '/FAQ',
+  shared: '/shared',
+  folder: '/folder',
 };
 
 export default routes;
